refactor(ui): move measured text ellipsis into shared primitives
`ellipsize_to_fit` lived inside the help overlay's private search module, so
no other surface could reach it and the tree already carried three near-copies
of measured trimming. Move it (with its ellipsis constant) next to
`text_extents_for` in `ui::primitives`, the shared text-measurement home, and
re-export it from `help_overlay::search` so the overlay's own call sites stay
unchanged.

No behavior change.
